test(api-core): move out more site-explorer tests
Refactor site-explorer integration tests out of api-core into the
site-explorer crate, using the shared test harness and reusable
environment helpers. Add configurable resource pools to the test
harness so migrated tests can provision secondary VTEP IPs without
api-core fixtures.
