+++ crates/api-core/src/tests/site_explorer.rs
@@ -20,31 +20,24 @@ use std::net::IpAddr;
 use std::str::FromStr;
 use std::sync::Arc;
 
-use carbide_site_explorer::config::{SiteExplorerConfig, SiteExplorerExploreMode};
+use carbide_site_explorer::config::SiteExplorerConfig;
 use carbide_site_explorer::{SiteExplorer, endpoint_exploration_work_key};
 use common::api_fixtures::TestEnv;
 use common::api_fixtures::endpoint_explorer::MockEndpointExplorer;
-use db::sku::CURRENT_SKU_VERSION;
-use db::{self, ObjectColumnFilter, ObjectFilter, explored_endpoints as db_explored_endpoints};
+use db::{self, ObjectColumnFilter, explored_endpoints as db_explored_endpoints};
 use ipnetwork::IpNetwork;
-use itertools::Itertools;
 use mac_address::MacAddress;
 use model::expected_machine::{ExpectedMachine, ExpectedMachineData};
 use model::hardware_info::HardwareInfo;
-use model::machine::machine_search_config::MachineSearchConfig;
-use model::machine::{LoadSnapshotOptions, Machine, ManagedHostStateSnapshot};
+use model::machine::ManagedHostStateSnapshot;
 use model::metadata::Metadata;
 use model::site_explorer::{
     Chassis, ComputerSystem, EndpointExplorationError, EndpointExplorationReport, EndpointType,
-    ExploredDpu, ExploredEndpoint, ExploredManagedHost, PreingestionState, UefiDevicePath,
+    ExploredEndpoint,
 };
 use model::switch::SwitchSearchFilter;
 use model::test_support::{DpuConfig, ManagedHostConfig};
-use rpc::forge::GetSiteExplorationRequest;
 use rpc::forge::forge_server::Forge;
-use rpc::site_explorer::{
-    ExploredDpu as RpcExploredDpu, ExploredManagedHost as RpcExploredManagedHost,
-};
 use rpc::{DiscoveryData, DiscoveryInfo, MachineDiscoveryInfo};
 use sqlx::PgPool;
 use tonic::Request;
@@ -63,85 +56,7 @@ use crate::tests::common::api_fixtures::network_segment::{
 use crate::tests::common::api_fixtures::site_explorer::MockExploredHost;
 use crate::tests::common::rpc_builder::DhcpDiscovery;
 
-#[derive(Clone, Debug)]
-struct FakeMachine {
-    pub mac: MacAddress,
-    pub dhcp_vendor: String,
-    pub relay_address: &'static str,
-    pub ip: String,
-}
-
 const UNDERLAY_RELAY: &str = "192.0.1.1";
-const ADMIN_RELAY: &str = "192.0.2.1";
-
-impl FakeMachine {
-    fn new_admin(mac: &str, vendor: &str) -> Self {
-        Self {
-            mac: mac.parse().unwrap(),
-            dhcp_vendor: vendor.to_string(),
-            relay_address: ADMIN_RELAY,
-            ip: String::new(),
-        }
-    }
-
-    fn new(mac: &str, vendor: &str) -> Self {
-        Self {
-            mac: mac.parse().unwrap(),
-            dhcp_vendor: vendor.to_string(),
-            relay_address: UNDERLAY_RELAY,
-            ip: String::new(),
-        }
-    }
-
-    fn as_mock_dpu(&self) -> DpuConfig {
-        DpuConfig {
-            bmc_mac_address: self.mac,
-            ..DpuConfig::default()
-        }
-    }
-
-    fn as_mock_host(&self, dpus: Vec<DpuConfig>) -> ManagedHostConfig {
-        ManagedHostConfig {
-            bmc_mac_address: self.mac,
-            dpus,
-            ..ManagedHostConfig::default()
-        }
-    }
-}
-
-trait DiscoverDhcp {
-    async fn discover_dhcp(&mut self, env: &TestEnv) -> Result<(), Box<dyn std::error::Error>>;
-}
-
-impl DiscoverDhcp for FakeMachine {
-    async fn discover_dhcp(&mut self, env: &TestEnv) -> Result<(), Box<dyn std::error::Error>> {
-        let response = env
-            .api
-            .discover_dhcp(
-                DhcpDiscovery::builder(self.mac, self.relay_address)
-                    .vendor_string(&self.dhcp_vendor)
-                    .tonic_request(),
-            )
-            .await?
-            .into_inner();
-        tracing::info!(
-            "DHCP with mac {} assigned ip {}",
-            self.mac,
-            response.address
-        );
-        self.ip = response.address;
-        Ok(())
-    }
-}
-
-impl DiscoverDhcp for Vec<FakeMachine> {
-    async fn discover_dhcp(&mut self, env: &TestEnv) -> Result<(), Box<dyn std::error::Error>> {
-        for machine in self.iter_mut() {
-            machine.discover_dhcp(env).await?
-        }
-        Ok(())
-    }
-}
 
 trait SiteExplorerConstructor {
     fn new_site_explorer(
@@ -169,1189 +84,6 @@ impl SiteExplorerConstructor for TestEnv {
             self.test_credential_manager.clone(),
         )
     }
-}
-
-#[sqlx_test]
-async fn test_site_explorer_default_pause_ingestion_and_poweron(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-    let underlay_segment = env.underlay_segment.unwrap();
-
-    let bmc_mac_address = "6a:6b:6c:6d:6e:6f".parse().unwrap();
-    let mut txn = pool.begin().await?;
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address,
-            data: ExpectedMachineData {
-                bmc_username: "ADMIN".into(),
-                bmc_password: "Pwd2023x0x0x0x0x7".into(),
-                serial_number: "VVG121GL".into(),
-                dpu_mode: model::expected_machine::DpuMode::NoDpu,
-                default_pause_ingestion_and_poweron: Some(true),
-                ..Default::default()
-            },
-        },
-    )
-    .await
-    .unwrap();
-    txn.commit().await?;
-
-    let mut machines = vec![FakeMachine::new(&bmc_mac_address.to_string(), "Vendor1")];
-    machines.discover_dhcp(&env).await?;
-
-    let mut txn = env.pool.begin().await?;
-    assert_eq!(
-        db::machine_interface::count_by_segment_id(&mut txn, &underlay_segment)
-            .await
-            .unwrap(),
-        1
-    );
-
-    let endpoint_explorer = Arc::new(MockEndpointExplorer::default());
-    let mock_host = machines[0].as_mock_host(vec![]);
-
-    endpoint_explorer.insert_endpoint_results(vec![(
-        machines[0].ip.parse().unwrap(),
-        Ok(mock_host.clone().into()),
-    )]);
-
-    let explorer_config = SiteExplorerConfig {
-        enabled: Arc::new(true.into()),
-        explorations_per_run: 2,
-        concurrent_explorations: 1,
-        run_interval: std::time::Duration::from_secs(1),
-        create_machines: Arc::new(true.into()),
-        ..Default::default()
-    };
-    let explorer = env.new_site_explorer(explorer_config, &endpoint_explorer);
-
-    // check the ingestion state of the machine
-    let response = env
-        .api
-        .determine_machine_ingestion_state(tonic::Request::new(rpc::forge::BmcEndpointRequest {
-            mac_address: Some("6a:6b:6c:6d:6e:6f".to_string()),
-            ip_address: "".to_string(),
-        }))
-        .await?;
-    assert_eq!(
-        rpc::forge::MachineIngestionState::NotDiscovered,
-        response.into_inner().machine_ingestion_state()
-    );
-
-    // run the exploration cycle
-    explorer.run_single_iteration().await.unwrap();
-
-    let mut txn = env.pool.begin().await?;
-    let explored = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-    assert_eq!(explored.len(), 1);
-    assert!(explored[0].pause_ingestion_and_poweron);
-
-    // make sure the machine has not been ingested
-    let response = env
-        .api
-        .determine_machine_ingestion_state(tonic::Request::new(rpc::forge::BmcEndpointRequest {
-            mac_address: Some("6a:6b:6c:6d:6e:6f".to_string()),
-            ip_address: "".to_string(),
-        }))
-        .await?;
-    assert_eq!(
-        rpc::forge::MachineIngestionState::WaitingForIngestion,
-        response.into_inner().machine_ingestion_state()
-    );
-
-    // now that the explored endpoint has been added to the DB, mark it as preingestion complete
-    db::explored_endpoints::set_preingestion_complete(explored[0].address, &mut txn)
-        .await
-        .unwrap();
-    txn.commit().await?;
-
-    // and run another exploration cycle
-    explorer.run_single_iteration().await.unwrap();
-
-    // make sure the machie still has not been ingested
-    let response = env
-        .api
-        .determine_machine_ingestion_state(tonic::Request::new(rpc::forge::BmcEndpointRequest {
-            mac_address: Some("6a:6b:6c:6d:6e:6f".to_string()),
-            ip_address: "".to_string(),
-        }))
-        .await?;
-    assert_eq!(
-        rpc::forge::MachineIngestionState::WaitingForIngestion,
-        response.into_inner().machine_ingestion_state()
-    );
-
-    let machine_snapshots =
-        db::managed_host::load_all(&env.pool, LoadSnapshotOptions::default()).await?;
-    assert_eq!(machine_snapshots.len(), 0);
-    let explored_managed_hosts = db::explored_managed_host::find_all(&env.pool).await?;
-    assert_eq!(explored_managed_hosts.len(), 0);
-
-    // now flip the flag and run another interation
-    let _ = env
-        .api
-        .allow_ingestion_and_power_on(tonic::Request::new(rpc::forge::BmcEndpointRequest {
-            mac_address: Some("6a:6b:6c:6d:6e:6f".to_string()),
-            ip_address: "".to_string(),
-        }))
-        .await?;
-
-    // run the exploration cycle
-    explorer.run_single_iteration().await.unwrap();
-
-    // the machine should be ingested now
-    // unfortunately, there is no way to test a hypothetical situation when
-    // an explored managed host has been created, but the machine has not
-    // been created yet as those are performed in the same site explorer
-    // iteration
-    let response = env
-        .api
-        .determine_machine_ingestion_state(tonic::Request::new(rpc::forge::BmcEndpointRequest {
-            mac_address: Some("6a:6b:6c:6d:6e:6f".to_string()),
-            ip_address: "".to_string(),
-        }))
-        .await?;
-    assert_eq!(
-        rpc::forge::MachineIngestionState::IngestionMachineCreated,
-        response.into_inner().machine_ingestion_state()
-    );
-
-    let explored_managed_hosts = db::explored_managed_host::find_all(&env.pool).await?;
-    assert_eq!(explored_managed_hosts.len(), 1);
-    let machine_snapshots =
-        db::managed_host::load_all(&env.pool, LoadSnapshotOptions::default()).await?;
-    assert_eq!(machine_snapshots.len(), 1);
-
-    Ok(())
-}
-
-#[sqlx_test]
-async fn test_site_explorer_main(pool: PgPool) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-    let underlay_segment = env.underlay_segment.unwrap();
-
-    // Let's create 3 machines on the underlay, and 1 on the admin network
-    // The 1 on the admin network is not supposed to be searched. This is verified
-    // by providing no mocked exploration data for this machine, which would lead
-    // to a panic if the machine is queried
-    let mut machines = vec![
-        // machines[0] is a DPU belonging to machines[1]
-        FakeMachine::new("B8:3F:D2:90:97:A6", "Vendor1"),
-        // machines[1] has 1 dpu (machines[0])
-        FakeMachine::new("AA:AB:AC:AD:AA:02", "Vendor2"),
-        // machines[2] has no DPUs
-        FakeMachine::new("AA:AB:AC:AD:AA:03", "Vendor3"),
-        // machines[3] is not on the underlay network and should not be searched.
-        FakeMachine::new_admin("AA:AB:AC:AD:BB:01", "VendorInvalidSegment"),
-    ];
-    machines.discover_dhcp(&env).await?;
-
-    let mut txn = env.pool.begin().await?;
-    assert_eq!(
-        db::machine_interface::count_by_segment_id(&mut txn, &underlay_segment)
-            .await
-            .unwrap(),
-        3
-    );
-    assert_eq!(
-        db::machine_interface::count_by_segment_id(&mut txn, env.admin_segment_ref())
-            .await
-            .unwrap(),
-        1
-    );
-    txn.commit().await.unwrap();
-
-    // Register `expected_machines` so site-explorer accepts these hosts: the
-    // host with a DPU pair takes the default `DpuMode`, and the zero-DPU host
-    // declares `NoDpu` to pass the strict ingestion gate.
-    let mut txn = env.pool.begin().await?;
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: machines[1].mac,
-            data: ExpectedMachineData {
-                serial_number: "host-with-dpu".to_string(),
-                ..Default::default()
-            },
-        },
-    )
-    .await?;
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: machines[2].mac,
-            data: ExpectedMachineData {
-                serial_number: "host-with-no-dpu".to_string(),
-                dpu_mode: model::expected_machine::DpuMode::NoDpu,
-                ..Default::default()
-            },
-        },
-    )
-    .await?;
-    txn.commit().await?;
-
-    let endpoint_explorer = Arc::new(MockEndpointExplorer::default());
-    let mock_dpu = machines[0].as_mock_dpu();
-
-    endpoint_explorer.insert_endpoint_results(vec![
-        (machines[0].ip.parse().unwrap(), Ok(mock_dpu.clone().into())),
-        (
-            machines[1].ip.parse().unwrap(),
-            Err(EndpointExplorationError::Unauthorized {
-                details: "Not authorized".to_string(),
-                response_body: None,
-                response_code: None,
-            }),
-        ),
-        (
-            machines[2].ip.parse().unwrap(),
-            Ok(EndpointExplorationReport {
-                endpoint_type: EndpointType::Bmc,
-                last_exploration_error: None,
-                last_exploration_latency: None,
-                vendor: Some(bmc_vendor::BMCVendor::Lenovo),
-                machine_id: None,
-                managers: Vec::new(),
-                systems: vec![ComputerSystem {
-                    serial_number: Some("0123456789".to_string()),
-                    ..Default::default()
-                }],
-                chassis: Vec::new(),
-                service: Vec::new(),
-                versions: HashMap::default(),
-                model: None,
-                machine_setup_status: None,
-                secure_boot_status: None,
-                lockdown_status: None,
-                power_shelf_id: None,
-                switch_id: None,
-                compute_tray_index: None,
-                physical_slot_number: None,
-                revision_id: None,
-                topology_id: None,
-                remediation_error: None,
-            }),
-        ),
-    ]);
-
-    let explorer_config = SiteExplorerConfig {
-        enabled: Arc::new(true.into()),
-        explorations_per_run: 2,
-        concurrent_explorations: 1,
-        run_interval: std::time::Duration::from_secs(1),
-        create_machines: Arc::new(true.into()),
-        create_power_shelves: Arc::new(true.into()),
-        explore_power_shelves_from_static_ip: Arc::new(true.into()),
-        power_shelves_created_per_run: 1,
-        create_switches: Arc::new(true.into()),
-        switches_created_per_run: 1,
-        ..Default::default()
-    };
-    let explorer = env.new_site_explorer(explorer_config, &endpoint_explorer);
-    let test_meter = &env.test_meter;
-
-    explorer.run_single_iteration().await.unwrap();
-    // Since we configured a limit of 2 entries, we should have those 2 results now
-    let mut txn = env.pool.begin().await?;
-    let explored = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-    txn.commit().await?;
-    assert_eq!(explored.len(), 2);
-
-    for report in &explored {
-        assert_eq!(report.report_version.version_nr(), 1);
-        let guard = endpoint_explorer.reports.lock().unwrap();
-        let res = guard.get(&report.address).unwrap().as_ref();
-        if res.is_err() {
-            assert_eq!(
-                res.unwrap_err(),
-                report.report.last_exploration_error.as_ref().unwrap()
-            );
-        } else {
-            assert_eq!(res.unwrap().endpoint_type, report.report.endpoint_type);
-            assert_eq!(res.unwrap().vendor, report.report.vendor);
-            assert_eq!(res.unwrap().managers, report.report.managers);
-            assert_eq!(res.unwrap().systems, report.report.systems);
-            assert_eq!(res.unwrap().chassis, report.report.chassis);
-            assert_eq!(res.unwrap().service, report.report.service);
-        }
-    }
-
-    // Retrieve the report via gRPC
-    let report = fetch_exploration_report(&env).await;
-    assert!(report.managed_hosts.is_empty());
-
-    // We should also have metric entries
-    assert_eq!(
-        test_meter
-            .formatted_metric("carbide_endpoint_explorations_count")
-            .unwrap(),
-        "2"
-    );
-    assert!(
-        test_meter
-            .formatted_metric("carbide_endpoint_exploration_success_count")
-            .is_some()
-    );
-    // The failure metric is not emitted if no failure happened
-    assert_eq!(
-        test_meter
-            .formatted_metric("carbide_endpoint_exploration_duration_milliseconds_count")
-            .unwrap_or("2".to_string()),
-        "2"
-    );
-    assert_eq!(
-        test_meter
-            .formatted_metric("carbide_site_exploration_identified_managed_hosts_count")
-            .unwrap(),
-        "0"
-    );
-    assert_eq!(
-        test_meter
-            .formatted_metric("carbide_site_explorer_created_machines_count")
-            .unwrap(),
-        "0"
-    );
-
-    // Running again should yield all 3 entries
-    explorer.run_single_iteration().await.unwrap();
-    // Since we configured a limit of 2 entries, we should have those 2 results now
-    let mut txn = env.pool.begin().await?;
-    let explored = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-    txn.commit().await?;
-    assert_eq!(explored.len(), 3);
-    let mut versions = Vec::new();
-    for report in &explored {
-        versions.push(report.report_version.version_nr());
-        let guard = endpoint_explorer.reports.lock().unwrap();
-        let res = guard.get(&report.address).unwrap().as_ref();
-        if res.is_err() {
-            assert_eq!(
-                res.unwrap_err(),
-                report.report.last_exploration_error.as_ref().unwrap()
-            );
-        } else {
-            assert_eq!(res.unwrap().endpoint_type, report.report.endpoint_type);
-            assert_eq!(res.unwrap().vendor, report.report.vendor);
-            assert_eq!(res.unwrap().managers, report.report.managers);
-            assert_eq!(res.unwrap().systems, report.report.systems);
-            assert_eq!(res.unwrap().chassis, report.report.chassis);
-            assert_eq!(res.unwrap().service, report.report.service);
-        }
-    }
-    versions.sort();
-    assert_eq!(&versions, &[1, 1, 2]);
-
-    // Retrieve the report via gRPC
-    let report = fetch_exploration_report(&env).await;
-    assert!(report.managed_hosts.is_empty());
-
-    assert_eq!(
-        test_meter
-            .formatted_metric("carbide_endpoint_explorations_count")
-            .unwrap(),
-        "2"
-    );
-    assert!(
-        test_meter
-            .formatted_metric("carbide_endpoint_exploration_success_count")
-            .is_some()
-    );
-    assert_eq!(
-        test_meter
-            .formatted_metric("carbide_endpoint_exploration_duration_milliseconds_count")
-            .unwrap_or("4".to_string()),
-        "4"
-    );
-    assert_eq!(
-        test_meter
-            .formatted_metric("carbide_site_exploration_identified_managed_hosts_count")
-            .unwrap(),
-        "0"
-    );
-    assert_eq!(
-        test_meter
-            .formatted_metric("carbide_site_explorer_created_machines_count")
-            .unwrap(),
-        "0"
-    );
-
-    // Now make 1 previously existing endpoint unreachable and 1 previously unreachable
-    // endpoint reachable and show the managed host.
-    // Both changes should show up after 2 updates
-    endpoint_explorer.insert_endpoint_results(vec![
-        (
-            machines[0].ip.parse().unwrap(),
-            Err(EndpointExplorationError::Unreachable {
-                details: Some("test_unreachable_detail".to_string()),
-            }),
-        ),
-        (
-            machines[1].ip.parse().unwrap(),
-            Ok(machines[1].as_mock_host(vec![mock_dpu.clone()]).into()),
-        ),
-    ]);
-
-    // We don't want to test the preingestion stuff here, so fake that it all completed successfully.
-    let mut txn = pool.begin().await?;
-    for addr in ["192.0.1.3", "192.0.1.4", "192.0.1.5"] {
-        db::explored_endpoints::set_preingestion_complete(
-            std::net::IpAddr::from_str(addr).unwrap(),
-            &mut txn,
-        )
-        .await
-        .unwrap();
-    }
-    txn.commit().await?;
-
-    explorer.run_single_iteration().await.unwrap();
-    explorer.run_single_iteration().await.unwrap();
-    let mut txn = env.pool.begin().await?;
-    let explored = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-    assert_eq!(explored.len(), 3);
-    let mut versions = Vec::new();
-    for report in &explored {
-        versions.push(report.report_version.version_nr());
-        assert_eq!(report.report.endpoint_type, EndpointType::Bmc);
-        match report.address.to_string() {
-            a if a == machines[0].ip => {
-                // The original successful report is retained, while only the latest
-                // exploration failure details are updated.
-                assert_eq!(report.report.vendor, Some(bmc_vendor::BMCVendor::Nvidia));
-                assert_eq!(
-                    report.report.last_exploration_error.clone().unwrap(),
-                    EndpointExplorationError::Unreachable {
-                        details: Some("test_unreachable_detail".to_string())
-                    }
-                );
-                assert!(report.report.last_exploration_latency.is_some());
-            }
-            a if a == machines[1].ip => {
-                assert_eq!(report.report.vendor, Some(bmc_vendor::BMCVendor::Dell));
-                assert!(report.report.last_exploration_error.is_none());
-            }
-            a if a == machines[2].ip => {
-                assert_eq!(report.report.vendor, Some(bmc_vendor::BMCVendor::Lenovo));
-                assert!(report.report.last_exploration_error.is_none());
-            }
-            _ => panic!("No other endpoints should be discovered"),
-        }
-    }
-    versions.sort();
-    // We run 4 iterations, which is enough for 8 machine scans
-    // => 2 Machines should have been scanned 3 times, and one 2 times
-    assert_eq!(&versions, &[2, 3, 3]);
-
-    let report = fetch_exploration_report(&env).await;
-    assert_eq!(report.endpoints.len(), 3);
-    let mut addresses: Vec<String> = report
-        .endpoints
-        .iter()
-        .map(|ep| ep.address.clone())
-        .collect();
-    addresses.sort();
-    let mut expected_addresses: Vec<String> = machines
-        .iter()
-        .filter(|m| m.relay_address == UNDERLAY_RELAY)
-        .map(|m| m.ip.to_string())
-        .collect();
-    expected_addresses.sort();
-    assert_eq!(addresses, expected_addresses);
-
-    // We should now have two managed hosts: One with a single DPU, and one with no DPUs.
-    assert_eq!(report.managed_hosts.len(), 2);
-    let managed_host_1 = report
-        .managed_hosts
-        .iter()
-        .find(|h| h.dpus.len() == 1)
-        .expect("Should have found one managed host with a single DPU")
-        .clone();
-    let managed_host_2 = report
-        .managed_hosts
-        .iter()
-        .find(|h| h.dpus.is_empty())
-        .expect("Should have found one managed host with zero DPUs")
-        .clone();
-
-    assert_eq!(
-        managed_host_1,
-        RpcExploredManagedHost {
-            host_bmc_ip: machines[1].ip.clone(),
-            dpu_bmc_ip: machines[0].ip.clone(),
-            host_pf_mac_address: Some(mock_dpu.host_mac_address.to_string()),
-            dpus: vec![RpcExploredDpu {
-                bmc_ip: machines[0].ip.clone(),
-                host_pf_mac_address: Some(mock_dpu.host_mac_address.to_string()),
-            }]
-        }
-    );
-
-    assert_eq!(
-        managed_host_2,
-        RpcExploredManagedHost {
-            host_bmc_ip: machines[2].ip.clone(),
-            dpu_bmc_ip: "".to_string(),
-            host_pf_mac_address: None,
-            dpus: vec![],
-        }
-    );
-
-    assert_eq!(
-        test_meter
-            .formatted_metric("carbide_site_exploration_identified_managed_hosts_count")
-            .unwrap(),
-        "2"
-    );
-
-    txn.commit().await?;
-    Ok(())
-}
-
-#[sqlx_test]
-async fn test_site_explorer_audit_exploration_results(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-    let underlay_segment = env.underlay_segment.unwrap();
-
-    let mut txn = pool.begin().await?;
-    for (bmc_mac_address, serial_number, fallback_dpu_serial_numbers) in [
-        ("0a:0b:0c:0d:0e:0f", "VVG121GG", vec![]),
-        ("1a:1b:1c:1d:1e:1f", "VVG121GH", vec![]),
-        ("2a:2b:2c:2d:2e:2f", "VVG121GI", vec![]),
-        ("3a:3b:3c:3d:3e:3f", "VVG121GJ", vec!["dpu_serial1"]),
-        (
-            "4a:4b:4c:4d:4e:4f",
-            "VVG121GK",
-            vec!["dpu_serial2", "dpu_serial3"],
-        ),
-        ("5a:5b:5c:5d:5e:5f", "VVG121GL", vec![]),
-    ] {
-        db::expected_machine::create(
-            &mut txn,
-            ExpectedMachine {
-                id: None,
-                bmc_mac_address: bmc_mac_address.parse().unwrap(),
-                data: ExpectedMachineData {
-                    bmc_username: "ADMIN".into(),
-                    bmc_password: "Pwd2023x0x0x0x0x7".into(),
-                    serial_number: serial_number.into(),
-                    fallback_dpu_serial_numbers: fallback_dpu_serial_numbers
-                        .into_iter()
-                        .map(ToString::to_string)
-                        .collect(),
-                    ..Default::default()
-                },
-            },
-        )
-        .await
-        .unwrap();
-    }
-    txn.commit().await?;
-
-    let mut machines = vec![
-        // This will be our expected DPU, and it will have the
-        // expected serial number, but we assume no DPUs are expected,
-        // should it still shouldn't be counted as `expected`        .
-        FakeMachine::new("5a:5b:5c:5d:5e:5f", "Vendor1"),
-        // This will be expected but unauthorized, and the serial is mismatched
-        FakeMachine::new("0a:0b:0c:0d:0e:0f", "Vendor3"),
-        // This host will be expected but missing credentials, and the serial is mismatched
-        FakeMachine::new("1a:1b:1c:1d:1e:1f", "Vendor3"),
-        // This host will be expected, but the serial number will be mismatched.
-        FakeMachine::new("2a:2b:2c:2d:2e:2f", "Vendor3"),
-        // This will be expected, with a good serial number.
-        // It will also have associated DPUs and should get a managed host.
-        FakeMachine::new("3a:3b:3c:3d:3e:3f", "Vendor3"),
-        // This host is not expected.
-        FakeMachine::new("ab:cd:ef:ab:cd:ef", "Vendor3"),
-        // This DPU is really not expected. (i.e. no DB entry)
-        FakeMachine::new("ef:cd:ab:ef:cd:ab", "Vendor3"),
-    ];
-
-    machines.discover_dhcp(&env).await?;
-
-    let mut txn = env.pool.begin().await?;
-    assert_eq!(
-        db::machine_interface::count_by_segment_id(&mut txn, &underlay_segment)
-            .await
-            .unwrap(),
-        7
-    );
-    txn.commit().await.unwrap();
-
-    // Make a mock host for machines[4] to generate the report
-    // This serial is from the create_expected_machine.sql seed.
-    let machine_4_host = ManagedHostConfig::with_serial("VVG121GJ".to_string());
-
-    let endpoint_explorer = Arc::new(MockEndpointExplorer::default());
-    endpoint_explorer.insert_endpoints(vec![
-        (
-            machines[0].ip.parse().unwrap(),
-            DpuConfig::with_serial("VVG121GL".to_string()).into(),
-        ),
-        (
-            machines[1].ip.parse().unwrap(),
-            EndpointExplorationReport {
-                endpoint_type: EndpointType::Bmc,
-                // Pretend there was previously a successful exploration
-                // but now something has gone wrong.
-                last_exploration_error: Some(EndpointExplorationError::Unauthorized {
-                    details: "Not authorized".to_string(),
-                    response_body: None,
-                    response_code: None,
-                }),
-                last_exploration_latency: None,
-                vendor: Some(bmc_vendor::BMCVendor::Lenovo),
-                machine_id: None,
-                model: None,
-                managers: Vec::new(),
-                systems: Vec::new(),
-                chassis: Vec::new(),
-                service: Vec::new(),
-                versions: HashMap::default(),
-                machine_setup_status: None,
-                secure_boot_status: None,
-                lockdown_status: None,
-                power_shelf_id: None,
-                switch_id: None,
-                compute_tray_index: None,
-                physical_slot_number: None,
-                revision_id: None,
-                topology_id: None,
-                remediation_error: None,
-            },
-        ),
-        (
-            machines[2].ip.parse().unwrap(),
-            EndpointExplorationReport {
-                endpoint_type: EndpointType::Bmc,
-                // Pretend there was previously a successful exploration
-                // but now something has gone wrong.
-                last_exploration_error: Some(EndpointExplorationError::MissingCredentials {
-                    key: "some_cred".to_string(),
-                    cause: "it's not there!".to_string(),
-                }),
-                last_exploration_latency: None,
-                vendor: Some(bmc_vendor::BMCVendor::Lenovo),
-                machine_id: None,
-                model: None,
-                managers: Vec::new(),
-                systems: Vec::new(),
-                chassis: Vec::new(),
-                service: Vec::new(),
-                versions: HashMap::default(),
-                machine_setup_status: None,
-                secure_boot_status: None,
-                lockdown_status: None,
-                power_shelf_id: None,
-                switch_id: None,
-                compute_tray_index: None,
-                physical_slot_number: None,
-                revision_id: None,
-                topology_id: None,
-                remediation_error: None,
-            },
-        ),
-        (
-            machines[3].ip.parse().unwrap(),
-            EndpointExplorationReport {
-                endpoint_type: EndpointType::Bmc,
-                last_exploration_error: None,
-                last_exploration_latency: None,
-                vendor: Some(bmc_vendor::BMCVendor::Lenovo),
-                machine_id: None,
-                model: None,
-                managers: Vec::new(),
-                systems: Vec::new(),
-                chassis: Vec::new(),
-                service: Vec::new(),
-                versions: HashMap::default(),
-                machine_setup_status: None,
-                secure_boot_status: None,
-                lockdown_status: None,
-                power_shelf_id: None,
-                switch_id: None,
-                compute_tray_index: None,
-                physical_slot_number: None,
-                revision_id: None,
-                topology_id: None,
-                remediation_error: None,
-            },
-        ),
-        (
-            machines[4].ip.parse().unwrap(),
-            machine_4_host.clone().into(),
-        ),
-        (
-            machines[5].ip.parse().unwrap(),
-            EndpointExplorationReport {
-                endpoint_type: EndpointType::Bmc,
-                last_exploration_error: None,
-                last_exploration_latency: None,
-                vendor: Some(bmc_vendor::BMCVendor::Lenovo),
-                machine_id: None,
-                model: None,
-                managers: Vec::new(),
-                systems: Vec::new(),
-                chassis: Vec::new(),
-                service: Vec::new(),
-                versions: HashMap::default(),
-                machine_setup_status: None,
-                secure_boot_status: None,
-                lockdown_status: None,
-                power_shelf_id: None,
-                switch_id: None,
-                compute_tray_index: None,
-                physical_slot_number: None,
-                revision_id: None,
-                topology_id: None,
-                remediation_error: None,
-            },
-        ),
-        (
-            // This is the DPU from machines[4]
-            machines[6].ip.parse().unwrap(),
-            machine_4_host.dpus[0].clone().into(),
-        ),
-    ]);
-
-    let explorer_config = SiteExplorerConfig {
-        enabled: Arc::new(true.into()),
-        explorations_per_run: 7,
-        concurrent_explorations: 1,
-        run_interval: std::time::Duration::from_secs(1),
-        create_machines: Arc::new(true.into()),
-        machines_created_per_run: 1,
-        override_target_ip: None,
-        override_target_port: None,
-        allow_changing_bmc_proxy: None,
-        bmc_proxy: Arc::default(),
-        reset_rate_limit: chrono::Duration::hours(1),
-        admin_segment_type_non_dpu: Arc::new(false.into()),
-        allocate_secondary_vtep_ip: false,
-        create_power_shelves: Arc::new(true.into()),
-        explore_power_shelves_from_static_ip: Arc::new(true.into()),
-        power_shelves_created_per_run: 1,
-        create_switches: Arc::new(true.into()),
-        switches_created_per_run: 1,
-        rotate_switch_nvos_credentials: Arc::new(false.into()),
-        dpu_mode: None,
-        // Tests use MockEndpointExplorer. So this doesn't affect anything.
-        explore_mode: SiteExplorerExploreMode::NvRedfish,
-    };
-    let explorer = env.new_site_explorer(explorer_config, &endpoint_explorer);
-    let test_meter = &env.test_meter;
-
-    explorer.run_single_iteration().await.unwrap();
-    // carbide_endpoint_exploration_preingestions_incomplete_overall_count
-    let m: HashMap<String, String> = test_meter
-        .parsed_metrics("carbide_endpoint_exploration_preingestions_incomplete_overall_count")
-        .into_iter()
-        .collect();
-
-    assert!(!m.is_empty());
-    assert_eq!(
-        m.get("{expectation=\"na\",machine_type=\"dpu\"}").unwrap(),
-        "2"
-    );
-    assert_eq!(
-        m.get("{expectation=\"expected\",machine_type=\"host\"}")
-            .unwrap(),
-        "4" // 2 normal + 2 previously explored but in an error state
-    );
-    assert_eq!(
-        m.get("{expectation=\"unexpected\",machine_type=\"host\"}")
-            .unwrap(),
-        "1"
-    );
-
-    let mut txn = pool.begin().await?;
-    for final_octet in 2..10 {
-        db::explored_endpoints::set_preingestion_complete(
-            std::net::IpAddr::from(std::net::Ipv4Addr::new(192, 0, 1, final_octet)),
-            &mut txn,
-        )
-        .await
-        .unwrap();
-    }
-    txn.commit().await?;
-    explorer.run_single_iteration().await.unwrap();
-
-    let mut txn = env.pool.begin().await?;
-    let explored = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-    txn.commit().await?;
-    assert_eq!(explored.len(), 7);
-
-    for report in &explored {
-        assert_eq!(report.report_version.version_nr(), 2);
-        let guard = endpoint_explorer.reports.lock().unwrap();
-        let res = guard.get(&report.address).unwrap().as_ref();
-        if res.is_err() {
-            assert_eq!(
-                res.unwrap_err(),
-                report.report.last_exploration_error.as_ref().unwrap()
-            );
-        } else {
-            assert_eq!(res.unwrap().endpoint_type, report.report.endpoint_type);
-            assert_eq!(res.unwrap().vendor, report.report.vendor);
-            assert_eq!(res.unwrap().managers, report.report.managers);
-            assert_eq!(res.unwrap().systems, report.report.systems);
-            assert_eq!(res.unwrap().chassis, report.report.chassis);
-            assert_eq!(res.unwrap().service, report.report.service);
-        }
-    }
-
-    // Retrieve the report via gRPC
-    let report = fetch_exploration_report(&env).await;
-
-    // We should have at least one managed host built by this point.
-    assert!(!report.managed_hosts.is_empty());
-
-    // Check for the expected metrics
-
-    // carbide_endpoint_exploration_failures_overall_count
-    let m: HashMap<String, String> = test_meter
-        .parsed_metrics("carbide_endpoint_exploration_failures_overall_count")
-        .into_iter()
-        .collect();
-
-    assert!(!m.is_empty());
-    assert!(m.get("{failure=\"unauthorized\"}").unwrap() == "1");
-    assert!(m.get("{failure=\"missing_credentials\"}").unwrap() == "1");
-
-    // carbide_endpoint_exploration_preingestions_incomplete_overall_count
-    let m: HashMap<String, String> = test_meter
-        .parsed_metrics("carbide_endpoint_exploration_preingestions_incomplete_overall_count")
-        .into_iter()
-        .collect();
-    // Everything should be done with preingestion now.
-    assert!(m.is_empty());
-
-    // carbide_endpoint_exploration_expected_serial_number_mismatches_overall_count
-    let m: HashMap<String, String> = test_meter
-        .parsed_metrics(
-            "carbide_endpoint_exploration_expected_serial_number_mismatches_overall_count",
-        )
-        .into_iter()
-        .collect();
-
-    assert!(!m.is_empty());
-    assert_eq!(m.get("{machine_type=\"host\"}").unwrap(), "3");
-
-    // carbide_endpoint_exploration_machines_explored_overall_count
-    let m: HashMap<String, String> = test_meter
-        .parsed_metrics("carbide_endpoint_exploration_machines_explored_overall_count")
-        .into_iter()
-        .collect();
-
-    assert!(!m.is_empty());
-    assert_eq!(
-        m.get("{expectation=\"na\",machine_type=\"dpu\"}").unwrap(),
-        "2"
-    );
-    assert_eq!(
-        m.get("{expectation=\"expected\",machine_type=\"host\"}")
-            .unwrap(),
-        "4"
-    );
-    assert_eq!(
-        m.get("{expectation=\"unexpected\",machine_type=\"host\"}")
-            .unwrap(),
-        "1"
-    );
-
-    // carbide_endpoint_exploration_expected_machines_missing_overall_count
-    assert_eq!(
-        test_meter
-            .formatted_metric(
-                "carbide_endpoint_exploration_expected_machines_missing_overall_count"
-            )
-            .unwrap(),
-        "1"
-    );
-
-    // carbide_endpoint_exploration_identified_managed_hosts_overall_count
-    let m: HashMap<String, String> = test_meter
-        .parsed_metrics("carbide_endpoint_exploration_identified_managed_hosts_overall_count")
-        .into_iter()
-        .collect();
-
-    assert!(!m.is_empty());
-    assert_eq!(m.get("{expectation=\"expected\"}").unwrap(), "1");
-
-    Ok(())
-}
-
-#[sqlx_test]
-async fn test_site_explorer_reexplore(pool: PgPool) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-    let underlay_segment = env.underlay_segment.unwrap();
-
-    let mut machines = vec![
-        FakeMachine::new("B8:3F:D2:90:97:A6", "Vendor1"),
-        FakeMachine::new("AA:AB:AC:AD:AA:02", "Vendor2"),
-    ];
-
-    machines.discover_dhcp(&env).await?;
-
-    let mut txn = env.pool.begin().await?;
-    assert_eq!(
-        db::machine_interface::count_by_segment_id(&mut txn, &underlay_segment)
-            .await
-            .unwrap(),
-        2
-    );
-    txn.commit().await.unwrap();
-
-    let endpoint_explorer = Arc::new(MockEndpointExplorer::default());
-    endpoint_explorer.insert_endpoint_results(vec![
-        (
-            machines[0].ip.parse().unwrap(),
-            Ok(DpuConfig::default().into()),
-        ),
-        (
-            machines[1].ip.parse().unwrap(),
-            Err(EndpointExplorationError::Unauthorized {
-                details: "Not authorized".to_string(),
-                response_body: None,
-                response_code: None,
-            }),
-        ),
-    ]);
-
-    let explorer_config = SiteExplorerConfig {
-        enabled: Arc::new(true.into()),
-        explorations_per_run: 1,
-        concurrent_explorations: 1,
-        run_interval: std::time::Duration::from_secs(1),
-        create_machines: Arc::new(false.into()),
-        create_power_shelves: Arc::new(true.into()),
-        explore_power_shelves_from_static_ip: Arc::new(true.into()),
-        power_shelves_created_per_run: 1,
-        create_switches: Arc::new(true.into()),
-        switches_created_per_run: 1,
-        ..Default::default()
-    };
-
-    let explorer = env.new_site_explorer(explorer_config, &endpoint_explorer);
-
-    explorer.run_single_iteration().await.unwrap();
-    // Since we configured a limit of 1 entries, we should have 1 results now
-    let mut txn = env.pool.begin().await?;
-    let explored = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-    txn.commit().await?;
-    assert_eq!(explored.len(), 1);
-    let explored_ip = explored[0].address;
-
-    for report in &explored {
-        assert_eq!(report.report_version.version_nr(), 1);
-        assert!(!report.exploration_requested);
-    }
-
-    // Re-exploring the first endpoint should prioritize it while preserving
-    // routine capacity for another endpoint.
-    env.api
-        .re_explore_endpoint(tonic::Request::new(rpc::forge::ReExploreEndpointRequest {
-            ip_address: explored_ip.to_string(),
-            if_version_match: None,
-        }))
-        .await
-        .unwrap();
-
-    // Calling the API should set the `exploration_requested` flag on the endpoint
-    let mut txn = env.pool.begin().await?;
-    let explored = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-    txn.commit().await?;
-    for report in &explored {
-        assert!(report.exploration_requested);
-    }
-
-    // The 2nd iteration updates the priority endpoint and still uses the
-    // routine budget to discover another endpoint.
-    explorer.run_single_iteration().await.unwrap();
-    let mut txn = env.pool.begin().await?;
-    let explored = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-    txn.commit().await?;
-    assert_eq!(explored.len(), 2);
-
-    let reexplored = explored
-        .iter()
-        .find(|report| report.address == explored_ip)
-        .unwrap();
-    assert_eq!(reexplored.report_version.version_nr(), 2);
-    assert!(!reexplored.exploration_requested);
-    let current_version = reexplored.report_version;
-
-    // Using if_version_match with an incorrect version does nothing
-    let unexpected_version = current_version.increment();
-    let e = env
-        .api
-        .re_explore_endpoint(tonic::Request::new(rpc::forge::ReExploreEndpointRequest {
-            ip_address: explored_ip.to_string(),
-            if_version_match: Some(unexpected_version.version_string()),
-        }))
-        .await
-        .expect_err("Should fail due to invalid version");
-    assert_eq!(e.code(), tonic::Code::FailedPrecondition);
-    assert_eq!(
-        e.message(),
-        format!(
-            "An object of type explored_endpoint was intended to be modified did not have the expected version {}",
-            unexpected_version.version_string()
-        )
-    );
-
-    let mut txn = env.pool.begin().await?;
-    let explored = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-    txn.commit().await?;
-    for report in &explored {
-        assert!(!report.exploration_requested);
-    }
-
-    // Using if_version_match with correct version string does flag the endpoint again
-    env.api
-        .re_explore_endpoint(tonic::Request::new(rpc::forge::ReExploreEndpointRequest {
-            ip_address: explored_ip.to_string(),
-            if_version_match: Some(current_version.version_string()),
-        }))
-        .await
-        .unwrap()
-        .into_inner();
-
-    let mut txn = env.pool.begin().await?;
-    let explored = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-    txn.commit().await?;
-    let reexplored = explored
-        .iter()
-        .find(|report| report.address == explored_ip)
-        .unwrap();
-    assert!(reexplored.exploration_requested);
-
-    // 3rd iteration still yields the same two known endpoints.
-    explorer.run_single_iteration().await.unwrap();
-    let mut txn = env.pool.begin().await?;
-    let explored = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-    txn.commit().await?;
-    assert_eq!(explored.len(), 2);
-
-    Ok(())
-}
-
-#[sqlx_test]
-async fn test_site_explorer_clear_last_known_error(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool).await;
-    let mut txn = db::Transaction::begin(&env.pool).await?;
-    let ip_address = "192.168.1.1";
-    let bmc_ip: IpAddr = IpAddr::from_str(ip_address)?;
-    let last_error = Some(EndpointExplorationError::Unreachable {
-        details: Some("test_unreachable_detail".to_string()),
-    });
-
-    let mut dpu_report1: EndpointExplorationReport = DpuConfig {
-        last_exploration_error: last_error.clone(),
-        ..DpuConfig::default()
-    }
-    .into();
-    dpu_report1.generate_machine_id(false)?;
-
-    db::explored_endpoints::insert(bmc_ip, &dpu_report1, false, &mut txn).await?;
-    txn.commit().await?;
-
-    txn = db::Transaction::begin(&env.pool).await?;
-    let nodes = db::explored_endpoints::find_all_by_ip(bmc_ip, &mut txn).await?;
-    assert_eq!(nodes.len(), 1);
-    let node = nodes.first();
-    assert_eq!(node.unwrap().report.last_exploration_error, last_error);
-
-    env.api
-        .clear_site_exploration_error(Request::new(rpc::forge::ClearSiteExplorationErrorRequest {
-            ip_address: ip_address.to_string(),
-        }))
-        .await
-        .unwrap()
-        .into_inner();
-
-    let nodes = db::explored_endpoints::find_all_by_ip(bmc_ip, &mut txn).await?;
-    assert_eq!(nodes.len(), 1);
-    let node = nodes.first();
-    assert_eq!(node.unwrap().report.last_exploration_error, None);
-
-    Ok(())
-}
-
-/// Clearing the site exploration error should also lift a terminal preingestion
-/// `Failed` state back to `Initial`, so an operator can retry preingestion
-/// without force-deleting and rediscovering the endpoint.
-#[sqlx_test]
-async fn test_clear_error_resets_failed_preingestion(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool).await;
-    let mut txn = db::Transaction::begin(&env.pool).await?;
-    let ip_address = "192.168.1.2";
-    let bmc_ip: IpAddr = IpAddr::from_str(ip_address)?;
-
-    let mut report: EndpointExplorationReport = DpuConfig::default().into();
-    report.generate_machine_id(false)?;
-    db::explored_endpoints::insert(bmc_ip, &report, false, &mut txn).await?;
-
-    // Put the endpoint in the terminal Failed preingestion state, as a failed
-    // BMC time sync would.
-    db::explored_endpoints::set_preingestion_failed(
-        bmc_ip,
-        "BMC time synchronization failed after 3 reset attempts. Time difference exceeds 5 minutes threshold.".to_string(),
-        &mut txn,
-    )
-    .await?;
-    txn.commit().await?;
-
-    env.api
-        .clear_site_exploration_error(Request::new(rpc::forge::ClearSiteExplorationErrorRequest {
-            ip_address: ip_address.to_string(),
-        }))
-        .await
-        .unwrap()
-        .into_inner();
-
-    let mut txn = db::Transaction::begin(&env.pool).await?;
-    let nodes = db::explored_endpoints::find_all_by_ip(bmc_ip, &mut txn).await?;
-    assert_eq!(nodes.len(), 1);
-    assert_eq!(
-        nodes.first().unwrap().preingestion_state,
-        PreingestionState::Initial,
-        "clearing the error should reset a Failed preingestion to Initial"
-    );
-
-    Ok(())
 }
 
 // Test that discover_machines will reject request of machine that was not created by site-explorer when create_machines = true
@@ -1409,195 +141,6 @@ async fn test_disable_machine_creation_outside_site_explorer(
 
     // assert!(dm_response.is_err_and(|e| e.message().contains("was not discovered by site-explore")));
 
-    Ok(())
-}
-
-#[sqlx_test]
-async fn test_fallback_dpu_serial(pool: PgPool) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    const HOST1_DPU_BMC_MAC: &str = "B8:3F:D2:90:97:A6";
-    const HOST1_BMC_MAC: &str = "AA:AB:AC:AD:AA:02";
-    const HOST1_DPU_SERIAL_NUMBER: &str = "host1_dpu_serial_number";
-
-    let mut host1_dpu_bmc = FakeMachine::new(HOST1_DPU_BMC_MAC, "NVIDIA/BF/BMC");
-
-    let mut host1_bmc = FakeMachine::new(HOST1_BMC_MAC, "Vendor2");
-
-    // Create dhcp entries and machine_interface entries for the machines
-    for machine in [&mut host1_dpu_bmc, &mut host1_bmc] {
-        machine.discover_dhcp(&env).await?;
-    }
-    let endpoint_explorer = Arc::new(MockEndpointExplorer::default());
-
-    // Create a host and dpu reports && host has no dpu_serial
-    let host1_dpu_report = DpuConfig {
-        serial: HOST1_DPU_SERIAL_NUMBER.to_string(),
-        bmc_mac_address: HOST1_DPU_BMC_MAC.parse()?,
-        ..DpuConfig::default()
-    };
-    let host1_report = ManagedHostConfig {
-        bmc_mac_address: HOST1_BMC_MAC.parse()?,
-        ..ManagedHostConfig::default()
-    };
-    endpoint_explorer.insert_endpoint_results(vec![
-        (
-            host1_dpu_bmc.ip.parse().unwrap(),
-            Ok(host1_dpu_report.into()),
-        ),
-        (host1_bmc.ip.parse().unwrap(), Ok(host1_report.into())),
-    ]);
-
-    let explorer_config = SiteExplorerConfig {
-        enabled: Arc::new(true.into()),
-        explorations_per_run: 10,
-        concurrent_explorations: 1,
-        run_interval: std::time::Duration::from_secs(1),
-        create_machines: Arc::new(true.into()),
-        allocate_secondary_vtep_ip: true,
-        create_power_shelves: Arc::new(true.into()),
-        explore_power_shelves_from_static_ip: Arc::new(true.into()),
-        power_shelves_created_per_run: 1,
-        create_switches: Arc::new(true.into()),
-        switches_created_per_run: 1,
-        ..Default::default()
-    };
-    let explorer = env.new_site_explorer(explorer_config, &endpoint_explorer);
-
-    // Create expected_machine entry for host1 w.o fallback_dpu_serial_number
-    let mut txn = env.pool.begin().await?;
-
-    // Create the SKU record first
-    let test_sku = model::sku::Sku {
-        schema_version: CURRENT_SKU_VERSION,
-        id: "Sku1".to_string(),
-        description: "Test SKU for site explorer test".to_string(),
-        created: chrono::Utc::now(),
-        components: model::sku::SkuComponents {
-            chassis: model::sku::SkuComponentChassis {
-                vendor: "Vendor1".to_string(),
-                model: "Chassis1".to_string(),
-                architecture: "x86_64".to_string(),
-            },
-            cpus: vec![],
-            gpus: vec![],
-            memory: vec![],
-            infiniband_devices: vec![],
-            storage: vec![],
-            tpm: None,
-        },
-        device_type: None, // This will result in "unknown" device type
-    };
-    db::sku::create(&mut txn, &test_sku).await?;
-
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: HOST1_BMC_MAC.to_string().parse().unwrap(),
-            data: ExpectedMachineData {
-                bmc_username: "user1".to_string(),
-                bmc_password: "pw".to_string(),
-                serial_number: "host1".to_string(),
-                fallback_dpu_serial_numbers: vec![],
-                metadata: Metadata::new_with_default_name(),
-                sku_id: Some("Sku1".to_string()),
-                default_pause_ingestion_and_poweron: None,
-                host_nics: vec![],
-                rack_id: None,
-                dpf_enabled: Some(true),
-                bmc_ip_address: None,
-                bmc_retain_credentials: None,
-                dpu_mode: Default::default(),
-                host_lifecycle_profile: Default::default(),
-            },
-        },
-    )
-    .await?;
-    txn.commit().await?;
-
-    // Run site explorer
-    explorer.run_single_iteration().await.unwrap();
-    let mut txn = env.pool.begin().await?;
-    let explored_endpoints = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-
-    // Mark explored endpoints as pre-ingestion_complete
-    for ee in &explored_endpoints {
-        db::explored_endpoints::set_preingestion_complete(ee.address, &mut txn).await?;
-    }
-    txn.commit().await?;
-
-    assert_eq!(explored_endpoints.len(), 2);
-
-    let mut explored_managed_hosts = db::explored_managed_host::find_all(&env.pool).await?;
-    let mut machines =
-        db::machine::find(&env.pool, ObjectFilter::All, MachineSearchConfig::default())
-            .await
-            .unwrap();
-
-    // There should be no managed host
-    assert_eq!(explored_managed_hosts.len(), 0);
-    assert_eq!(machines.len(), 0);
-
-    // Now update expected_machine entry with fallback_dpu_serial
-    let mut txn = env.pool.begin().await?;
-    let mut host1_expected_machine =
-        db::expected_machine::find_by_bmc_mac_address(txn.as_mut(), HOST1_BMC_MAC.parse().unwrap())
-            .await?
-            .expect("Expected machine not found");
-    host1_expected_machine.data = ExpectedMachineData {
-        bmc_username: "user1".to_string(),
-        bmc_password: "pw".to_string(),
-        serial_number: "host1".to_string(),
-        fallback_dpu_serial_numbers: vec![HOST1_DPU_SERIAL_NUMBER.to_string()],
-        metadata: Metadata::new_with_default_name(),
-        sku_id: None,
-        default_pause_ingestion_and_poweron: None,
-        host_nics: vec![],
-        rack_id: None,
-        dpf_enabled: Some(true),
-        bmc_ip_address: None,
-        bmc_retain_credentials: None,
-        dpu_mode: Default::default(),
-        host_lifecycle_profile: Default::default(),
-    };
-    db::expected_machine::update(&mut txn, &host1_expected_machine).await?;
-    txn.commit().await?;
-
-    explorer.run_single_iteration().await.unwrap();
-    explored_managed_hosts = db::explored_managed_host::find_all(&env.pool).await?;
-    machines = db::machine::find(&env.pool, ObjectFilter::All, MachineSearchConfig::default())
-        .await
-        .unwrap();
-
-    // We should see one explored_managed host && 2 machines
-    assert_eq!(
-        <Vec<ExploredManagedHost> as AsRef<Vec<ExploredManagedHost>>>::as_ref(
-            &explored_managed_hosts
-        )
-        .len(),
-        1
-    );
-    assert_eq!(
-        <Vec<Machine> as AsRef<Vec<Machine>>>::as_ref(&machines).len(),
-        2
-    );
-
-    // Make sure they are the machines we just created
-    let mut bmc_ip_addresses = vec![explored_managed_hosts[0].host_bmc_ip.to_string()];
-    for dpu in explored_managed_hosts[0].clone().dpus {
-        bmc_ip_addresses.push(dpu.bmc_ip.to_string())
-    }
-    assert_eq!(bmc_ip_addresses.len(), 2);
-    for bmc_ip in bmc_ip_addresses {
-        assert!(
-            <Vec<Machine> as AsRef<Vec<Machine>>>::as_ref(&machines)
-                .iter()
-                .any(|x| { x.bmc_info.ip.is_some_and(|ip| ip.to_string() == bmc_ip) })
-        );
-    }
     Ok(())
 }
 
@@ -1727,80 +270,6 @@ async fn test_site_explorer_health_report(pool: PgPool) -> Result<(), Box<dyn st
     Ok(())
 }
 
-async fn fetch_exploration_report(env: &TestEnv) -> rpc::site_explorer::SiteExplorationReport {
-    env.api
-        .get_site_exploration_report(tonic::Request::new(GetSiteExplorationRequest::default()))
-        .await
-        .unwrap()
-        .into_inner()
-}
-
-#[sqlx_test]
-async fn test_fetch_host_primary_interface_mac(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let mut mock_dpus = (0..NUM_DPUS).map(|_| DpuConfig::default()).collect_vec();
-
-    // Make the second DPU have the lower-numbered UEFI device path... we will assert later that
-    // it's the primary DPU.
-    mock_dpus[0].override_hosts_uefi_device_path = Some(
-        UefiDevicePath::from_str("PciRoot(0x8)/Pci(0x2,0xa)/Pci(0x1,0x1)/MAC(A088C208545C,0x1)")
-            .unwrap(),
-    );
-    mock_dpus[1].override_hosts_uefi_device_path = Some(
-        UefiDevicePath::from_str("PciRoot(0x8)/Pci(0x2,0xa)/Pci(0x0,0x2)/MAC(A088C208545C,0x1)")
-            .unwrap(),
-    );
-
-    let host_report: EndpointExplorationReport =
-        ManagedHostConfig::with_dpus(mock_dpus.clone()).into();
-
-    const NUM_DPUS: usize = 2;
-
-    let env = common::api_fixtures::create_test_env(pool).await;
-    let mut txn = env.pool.begin().await?;
-    let mut oob_interfaces = Vec::new();
-    let mut explored_dpus = Vec::new();
-
-    for (i, mock_dpu) in mock_dpus.iter().enumerate() {
-        let oob_mac = mock_dpu.bmc_mac_address;
-        let response = env
-            .api
-            .discover_dhcp(
-                DhcpDiscovery::builder(oob_mac, "192.0.1.1")
-                    .vendor_string("NVIDIA/BF/BMC")
-                    .tonic_request(),
-            )
-            .await
-            .unwrap()
-            .into_inner();
-
-        assert!(!response.address.is_empty());
-        let oob_interface =
-            db::machine_interface::find_by_mac_address(txn.as_mut(), oob_mac).await?;
-        assert!(oob_interface[0].primary_interface);
-        oob_interfaces.push(oob_interface[0].clone());
-
-        let mut dpu_report: EndpointExplorationReport = mock_dpu.clone().into();
-        dpu_report.generate_machine_id(false)?;
-        let dpu_report = Arc::new(dpu_report);
-        explored_dpus.push(ExploredDpu {
-            bmc_ip: IpAddr::from_str(format!("192.168.1.{i}").as_str())?,
-            host_pf_mac_address: Some(mock_dpu.host_mac_address),
-            report: dpu_report,
-        });
-    }
-
-    let expected_mac: MacAddress = mock_dpus[1].host_mac_address;
-    let mac = host_report
-        .fetch_host_primary_interface_mac(&explored_dpus)
-        .unwrap();
-    assert_eq!(mac, expected_mac);
-    Ok(())
-}
-
-/// Test the [`api_fixtures::site_explorer::new_host`] factory with various configurations and make
-/// sure they work.
 #[sqlx_test]
 async fn test_site_explorer_new_host_fixture(
     pool: PgPool,
@@ -2328,153 +797,6 @@ async fn test_delete_explored_endpoint(pool: PgPool) -> Result<(), Box<dyn std::
 }
 
 #[sqlx_test]
-async fn test_machine_creation_with_sku(pool: PgPool) -> Result<(), Box<dyn std::error::Error>> {
-    let env = common::api_fixtures::create_test_env(pool.clone()).await;
-
-    const HOST1_DPU_BMC_MAC: &str = "B8:3F:D2:90:97:A6";
-    const HOST1_BMC_MAC: &str = "AA:AB:AC:AD:AA:02";
-    const HOST1_DPU_SERIAL_NUMBER: &str = "host1_dpu_serial_number";
-
-    let mut host1_dpu_bmc = FakeMachine::new(HOST1_DPU_BMC_MAC, "NVIDIA/BF/BMC");
-
-    let mut host1_bmc = FakeMachine::new(HOST1_BMC_MAC, "Vendor2");
-
-    // Create dhcp entries and machine_interface entries for the machines
-    for machine in [&mut host1_dpu_bmc, &mut host1_bmc] {
-        machine.discover_dhcp(&env).await?;
-    }
-    let endpoint_explorer = Arc::new(MockEndpointExplorer::default());
-
-    // Create a host and dpu reports && host has no dpu_serial
-    let host1_dpu_report = DpuConfig {
-        serial: HOST1_DPU_SERIAL_NUMBER.to_string(),
-        bmc_mac_address: HOST1_DPU_BMC_MAC.parse()?,
-        ..DpuConfig::default()
-    };
-    let host1_report = ManagedHostConfig {
-        bmc_mac_address: HOST1_BMC_MAC.parse()?,
-        ..ManagedHostConfig::default()
-    };
-    endpoint_explorer.insert_endpoint_results(vec![
-        (
-            host1_dpu_bmc.ip.parse().unwrap(),
-            Ok(host1_dpu_report.into()),
-        ),
-        (host1_bmc.ip.parse().unwrap(), Ok(host1_report.into())),
-    ]);
-
-    let explorer_config = SiteExplorerConfig {
-        enabled: Arc::new(true.into()),
-        explorations_per_run: 10,
-        concurrent_explorations: 1,
-        run_interval: std::time::Duration::from_secs(1),
-        create_machines: Arc::new(true.into()),
-        allocate_secondary_vtep_ip: true,
-        create_power_shelves: Arc::new(true.into()),
-        explore_power_shelves_from_static_ip: Arc::new(true.into()),
-        power_shelves_created_per_run: 1,
-        create_switches: Arc::new(true.into()),
-        switches_created_per_run: 1,
-        ..Default::default()
-    };
-    let explorer = env.new_site_explorer(explorer_config, &endpoint_explorer);
-    let test_meter = &env.test_meter;
-
-    // Create expected_machine entry for host1 w.o fallback_dpu_serial_number
-    let mut txn = env.pool.begin().await?;
-
-    // Create the SKU record first
-    let test_sku = model::sku::Sku {
-        schema_version: CURRENT_SKU_VERSION,
-        id: "Sku1".to_string(),
-        description: "Test SKU for site explorer test".to_string(),
-        created: chrono::Utc::now(),
-        components: model::sku::SkuComponents {
-            chassis: model::sku::SkuComponentChassis {
-                vendor: "Vendor1".to_string(),
-                model: "Chassis1".to_string(),
-                architecture: "x86_64".to_string(),
-            },
-            cpus: vec![],
-            gpus: vec![],
-            memory: vec![],
-            infiniband_devices: vec![],
-            storage: vec![],
-            tpm: None,
-        },
-        device_type: None, // This will result in "unknown" device type
-    };
-    db::sku::create(&mut txn, &test_sku).await?;
-
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: HOST1_BMC_MAC.to_string().parse().unwrap(),
-            data: ExpectedMachineData {
-                bmc_username: "user1".to_string(),
-                bmc_password: "pw".to_string(),
-                serial_number: "host1".to_string(),
-                fallback_dpu_serial_numbers: vec![],
-                metadata: Metadata::new_with_default_name(),
-                sku_id: Some("Sku1".to_string()),
-                default_pause_ingestion_and_poweron: None,
-                host_nics: vec![],
-                rack_id: None,
-                dpf_enabled: Some(true),
-                bmc_ip_address: None,
-                bmc_retain_credentials: None,
-                dpu_mode: Default::default(),
-                host_lifecycle_profile: Default::default(),
-            },
-        },
-    )
-    .await?;
-    txn.commit().await?;
-
-    // Run site explorer
-    explorer.run_single_iteration().await.unwrap();
-    let mut txn = env.pool.begin().await?;
-    let explored_endpoints = db::explored_endpoints::find_all(txn.as_mut())
-        .await
-        .unwrap();
-
-    // Mark explored endpoints as pre-ingestion_complete
-    for ee in &explored_endpoints {
-        db::explored_endpoints::set_preingestion_complete(ee.address, &mut txn).await?;
-    }
-    txn.commit().await?;
-
-    assert_eq!(explored_endpoints.len(), 2);
-
-    let machines = db::machine::find(&env.pool, ObjectFilter::All, MachineSearchConfig::default())
-        .await
-        .unwrap();
-
-    for m in machines {
-        if m.is_dpu() {
-            assert_eq!(m.hw_sku, None);
-        } else {
-            assert_eq!(m.hw_sku, Some("Sku1".to_string()));
-            assert!(m.dpf.enabled);
-        }
-    }
-
-    // Verify expected machine SKU metrics
-    let expected_metrics: HashMap<String, String> = test_meter
-        .parsed_metrics("carbide_site_exploration_expected_machines_sku_count")
-        .into_iter()
-        .collect();
-
-    // We should have metrics for expected machines
-    assert!(!expected_metrics.is_empty());
-    // The SKU "Sku1" has device_type=None, so it should be counted with device_type="unknown"
-    assert!(expected_metrics.contains_key("{device_type=\"unknown\",sku_id=\"Sku1\"}"));
-
-    Ok(())
-}
-
-#[sqlx_test]
 async fn test_site_explorer_switch_discovery(
     pool: PgPool,
 ) -> Result<(), Box<dyn std::error::Error>> {
@@ -2693,85 +1015,6 @@ async fn test_get_machine_position_info_no_endpoint(
     assert_eq!(info.compute_tray_index, None);
     assert_eq!(info.topology_id, None);
     assert_eq!(info.revision_id, None);
-
-    Ok(())
-}
-
-/// Integration regression guard for the auto-correct path: when an
-/// `ExpectedMachine` declares `DpuMode::NicMode` but the discovered DPU
-/// hardware is reporting `nic_mode: Dpu`, site-explorer should call
-/// `set_nic_mode(Nic)` on the DPU during its per-host matching loop.
-///
-/// This exercises the full wire (site-explorer iteration → per-host mode
-/// resolution → `check_and_configure_dpu_mode` → mock Redfish
-/// `set_nic_mode`) that the unit tests only cover in pieces.
-#[sqlx_test]
-async fn test_site_explorer_auto_corrects_nic_mode_per_expected_machine(
-    pool: PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    use model::expected_machine::{DpuMode, ExpectedMachine, ExpectedMachineData};
-    use model::site_explorer::NicMode;
-
-    let env = common::api_fixtures::create_test_env(pool).await;
-
-    // DPU hardware reports DPU mode (so it looks like a "properly
-    // configured" DPU to the BF3-DPU heuristic) -- the operator-declared
-    // override is what forces the correction to NIC mode.
-    let dpu_config = DpuConfig {
-        nic_mode: Some(NicMode::Dpu),
-        ..DpuConfig::default()
-    };
-    let mock_host = model::test_support::ManagedHostConfig::with_dpus(vec![dpu_config.clone()]);
-    let host_bmc_mac = mock_host.bmc_mac_address;
-
-    // Seed an ExpectedMachine with `dpu_mode: NicMode` that matches the
-    // mock host's BMC MAC. Site-explorer's per-host resolution will look
-    // this up by IP via the expected-endpoint index after DHCP assigns
-    // the host its BMC IP.
-    let mut txn = env.pool.begin().await?;
-    db::expected_machine::create(
-        &mut txn,
-        ExpectedMachine {
-            id: None,
-            bmc_mac_address: host_bmc_mac,
-            data: ExpectedMachineData {
-                bmc_username: "ADMIN".to_string(),
-                bmc_password: "PASS".to_string(),
-                serial_number: "EM-866-NIC-OVERRIDE".to_string(),
-                metadata: model::metadata::Metadata::new_with_default_name(),
-                dpu_mode: DpuMode::NicMode,
-                ..Default::default()
-            },
-        },
-    )
-    .await?;
-    txn.commit().await?;
-
-    // Drive the same ingestion flow as the singledpu fixture test: BMC
-    // DHCP for host + DPU, seed mock exploration results, run site-
-    // explorer iteration. We don't care about the final managed host
-    // state here -- we only care that `set_nic_mode` was called with the
-    // right target during the matching loop.
-    common::api_fixtures::site_explorer::MockExploredHost::new(&env, mock_host)
-        .discover_dhcp_host_bmc(|_, _| Ok(()))
-        .await?
-        .discover_dhcp_dpu_bmc(0, |_, _| Ok(()))
-        .await?
-        .insert_site_exploration_results()?
-        // First iteration: initial endpoint exploration.
-        .run_site_explorer_iteration()
-        .await
-        .mark_preingestion_complete()
-        .await?
-        // Second iteration: per-host DPU matching + check_and_configure_dpu_mode.
-        .run_site_explorer_iteration()
-        .await;
-
-    let calls = env.endpoint_explorer.set_nic_mode_calls.lock().unwrap();
-    assert!(
-        calls.iter().any(|(_, mode)| *mode == NicMode::Nic),
-        "expected at least one set_nic_mode(Nic) call triggered by the operator's NicMode declaration; calls so far: {calls:?}"
-    );
 
     Ok(())
 }
