+++ crates/test-harness/src/resource_pool.rs
@@ -1,0 +1,99 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+use std::collections::HashMap;
+use std::iter;
+
+use ipnetwork::IpNetwork;
+use model::resource_pool;
+use model::resource_pool::common::SECONDARY_VTEP_IP;
+use model::resource_pool::{ResourcePoolDef, ResourcePoolType};
+
+#[derive(Default)]
+pub struct ResourcePoolBuilder {
+    secondary_vtep_ip: Option<IpNetwork>,
+}
+
+impl ResourcePoolBuilder {
+    pub fn with_secondary_vtep_ip(self, addr: &str) -> Self {
+        // This "allow" should be removed when more options will be
+        // added to the builder.
+        #[allow(clippy::needless_update)]
+        Self {
+            secondary_vtep_ip: Some(
+                addr.parse()
+                    .expect("correct IP address / mask must be provided"),
+            ),
+            ..self
+        }
+    }
+
+    pub fn build(self) -> HashMap<String, ResourcePoolDef> {
+        let int_range_pool = |ranges: &[(u32, u32)]| resource_pool::ResourcePoolDef {
+            pool_type: resource_pool::ResourcePoolType::Integer,
+            ranges: ranges
+                .iter()
+                .map(|(start, end)| resource_pool::Range {
+                    start: start.to_string(),
+                    end: end.to_string(),
+                    auto_assign: true,
+                })
+                .collect(),
+            prefix: None,
+            delegate_prefix_len: None,
+        };
+
+        [
+            (
+                resource_pool::common::LOOPBACK_IP.to_string(),
+                resource_pool::ResourcePoolDef {
+                    pool_type: resource_pool::ResourcePoolType::Ipv4,
+                    prefix: Some("172.20.0.0/24".to_string()),
+                    ranges: vec![],
+                    delegate_prefix_len: None,
+                },
+            ),
+            (
+                resource_pool::common::VLANID.to_string(),
+                int_range_pool(&[(1, 2)]),
+            ),
+            (
+                resource_pool::common::VNI.to_string(),
+                int_range_pool(&[(10001, 10002)]),
+            ),
+            (
+                resource_pool::common::VPC_VNI.to_string(),
+                int_range_pool(&[(20001, 20002), (60001, 60002)]),
+            ),
+        ]
+        .into_iter()
+        .map(Some)
+        .chain(iter::once(self.secondary_vtep_ip.map(|network| {
+            (
+                SECONDARY_VTEP_IP.to_string(),
+                ResourcePoolDef {
+                    pool_type: ResourcePoolType::Ipv4,
+                    prefix: Some(network.to_string()),
+                    ranges: vec![],
+                    delegate_prefix_len: None,
+                },
+            )
+        })))
+        .flatten()
+        .collect()
+    }
+}
